dual_system/requirements.txt
flask
python-dotenv
google-generativeai
openai
spacy
redis
psycopg2-binary
neo4j
flask-login
werkzeug